cleanup creating manager
